+++ packages/fit/src/adapters/extensions/extensions.extractor.ts
@@ -1,3 +1,4 @@
+import { extractFieldsFromMessageArray } from "./developer-fields.extractor";
 import { fitMessageKeySchema } from "../schemas/fit-message-keys";
 import type { FitMessages } from "../shared/types";
 import type { Logger } from "@kaiord/core";
@@ -36,23 +37,6 @@ const extractUnknownMessages = (
   return unknownMessages;
 };
 
-const extractFieldsFromMessageArray = (
-  messages: unknown[]
-): Array<Record<string, unknown>> => {
-  const developerFields: Array<Record<string, unknown>> = [];
-  for (const message of messages) {
-    if (message && typeof message === "object") {
-      const devFields = extractDeveloperFields(
-        message as Record<string, unknown>
-      );
-      if (devFields.length > 0) {
-        developerFields.push(...devFields);
-      }
-    }
-  }
-  return developerFields;
-};
-
 const extractAllDeveloperFields = (
   messages: FitMessages
 ): Array<Record<string, unknown>> => {
@@ -65,23 +49,6 @@ const extractAllDeveloperFields = (
   }
 
   return developerFields;
-};
-
-const extractDeveloperFields = (
-  message: Record<string, unknown>
-): Array<Record<string, unknown>> => {
-  const devFields: Array<Record<string, unknown>> = [];
-
-  for (const [key, value] of Object.entries(message)) {
-    if (key.startsWith("developer_") || key.includes("DeveloperField")) {
-      devFields.push({
-        fieldName: key,
-        value,
-      });
-    }
-  }
-
-  return devFields;
 };
 
 const buildExtensions = (
