+++ packages/tcx/src/adapters/workout/step.converter.ts
@@ -1,41 +1,7 @@
-import {
-  extractExtensions,
-  extractIntensity,
-  extractPowerFromExtensions,
-} from "./step-helpers";
+import { extractExtensions, extractIntensity } from "./step-helpers";
+import { convertTargetWithExtensions } from "./target-with-extensions.helper";
 import { convertTcxDuration } from "../duration/duration.mapper";
-import { convertTcxTarget } from "../target/target.mapper";
 import type { Duration, Logger, Target, WorkoutStep } from "@kaiord/core";
-
-const convertTargetWithExtensions = (
-  tcxStep: Record<string, unknown>,
-  extensions: Record<string, unknown> | undefined,
-  logger: Logger
-): Target | null => {
-  const target = convertTcxTarget(
-    tcxStep.Target as Record<string, unknown> | undefined,
-    logger
-  );
-  if (!target) return null;
-
-  if (target.type === "open" && extensions) {
-    const powerWatts = extractPowerFromExtensions(extensions, logger);
-    if (powerWatts !== undefined) {
-      logger.debug("Converting open target to power target from extensions", {
-        watts: powerWatts,
-      });
-      return {
-        type: "power",
-        value: {
-          unit: "watts",
-          value: powerWatts,
-        },
-      };
-    }
-  }
-
-  return target;
-};
 
 type BuildStepInput = {
   stepIndex: number;
